+++ contracts/core/wmoa-swap/tests/wmoa_swap_scenario_rs_test.rs
@@ -2,11 +2,11 @@ use dharitri_sc_scenario::*;
 
 fn world() -> ScenarioWorld {
     let mut blockchain = ScenarioWorld::new();
-    blockchain.set_current_dir_from_workspace("contracts/core/wegld-swap");
+    blockchain.set_current_dir_from_workspace("contracts/core/wmoa-swap");
 
     blockchain.register_contract(
-        "file:output/dharitri-wegld-swap-sc.wasm",
-        dharitri_wegld_swap_sc::ContractBuilder,
+        "file:output/dharitri-wmoa-swap-sc.wasm",
+        dharitri_wmoa_swap_sc::ContractBuilder,
     );
     blockchain
 }