+++ contracts/examples/multisig/tests/multisig_scenario_go_test.rs
@@ -67,8 +67,8 @@ fn interactor_nft_all_roles_go() {
 }
 
 #[test]
-fn interactor_wegld_go() {
-    world().run("scenarios/interactor_wegld.scen.json");
+fn interactor_wmoa_go() {
+    world().run("scenarios/interactor_wmoa.scen.json");
 }
 
 #[test]